MyQueue refactored
