+++ Fedya/FANUC Client/MyQueue.cpp
@@ -1,2 +1,1 @@
 #include "MyQueue.hpp"
-
