+++ Fedya/FANUC Client/MyQueue.hpp
@@ -7,122 +7,107 @@
 #include<queue>
 #include<mutex>
 
-//обртка над очередью
+/**
+ * \brief class which impliment a queue for multithreading and time measurement
+ * \tparam T object in queue
+ */
 template<typename T>
 class MyQueue
 {
-	std::queue<std::pair<int,T>> _q;
+	/**
+	 * \brief basic queue with elements
+	 */
+	std::queue<std::pair<clock_t,T>> _q;
+	
+	/**
+	 * \brief mutex for locking threads
+	 */
 	std::mutex _mt;
-	int _sizeOfQueue, _timeDifference,_interfaceId1,_interfaceId2;
+
+	/**
+	 * \brief size of queue for interface
+	 */
+	int _sizeOfQueue{ 0 };
+	
+	/**
+	 * \brief time differnce of last erased element betwen adding and erasing elemnt in queue
+	 */
+	int _timeDifference{ 0 };
+	
+	/**
+	 * \brief id of first interface field
+	 */
+	int _interfaceId1;
+	
+	/**
+	 * \brief if of secind interface id
+	 */
+	int _interfaceId2;
+	
+	/**
+	 * \brief flag if this queue witg interface or without
+	 */
 	bool _withInterface;
 public:
-	explicit MyQueue(std::string comment = "MyQueue")
-	{
-		_sizeOfQueue = 0;
-		_timeDifference = 0;
-		_withInterface = comment != "without interface";
-		if (_withInterface) 
-		{
-			_interfaceId1 = myInterface::MyShower::getInstance().addField(
-				myInterface::Field(std::string("size of ") + comment + ": ", &_sizeOfQueue));
-			_interfaceId2 = myInterface::MyShower::getInstance().addField(
-				myInterface::Field(std::string("time difference between inserting and erasing point in ") + comment + ": ", &_timeDifference));
-		}
-	}
+	/**
+	 * \brief contructor of queue
+	 * \param comment discription of this queue which will show in interface if it isn't specify this queue would be without interface
+	 */
+	explicit MyQueue(std::string comment = "without interface");
 
-	//метод добавления элемента в очередь
-	void push(T elem)
-	{
-		int t = clock();
-		_mt.lock();
-		_q.push(std::make_pair(t,elem));
-		_mt.unlock();
-		++_sizeOfQueue;
-	}
+	/**
+	 * \brief function for adding element
+	 * \param elem element for adding
+	 */
+	void push(T elem);
 
-	//метод прверки на пустоту
-	bool empty()
-	{
-		_mt.lock();
-		const bool ans = _q.empty();
-		_mt.unlock();
-		return ans;
-	}
+	/**
+	 * \brief fuction for checking if this queue is empty
+	 * \return return true if epmty, else false
+	 */
+	bool empty();
 
-	//метод получения размера
-	int size()
-	{
-		_mt.lock();
-		const int ans = _q.size();
-		_mt.unlock();
-		return ans;
-	}
+	/**
+	 * \brief function for checking size of queue
+	 * \return retrun a size of queue
+	 */
+	int size();
 
-	//метод получения временной отметки первой точки
-	int frontTime()
-	{
-		_mt.lock();
-		const int ans = _q.front().first;
-		_mt.unlock();
-		return ans;
-	}
+	/**
+	 * \brief function for getting time when front element was added
+	 * \return return time of adding front element
+	 */
+	clock_t frontTime();
 
-	//метод получения первого элемента
-	T front()
-	{
-		_mt.lock();
-		const T elem = _q.front().second;
-		_mt.unlock();
-		return elem;
-	}
+	/**
+	 * \brief function for getting front element
+	 * \return front element
+	 */
+	T front();
 
-	//метод удаления первого элемента
-	void pop()
-	{
-		_mt.lock();
-		_timeDifference = clock() - _q.front().first;
-		_q.pop();
-		--_sizeOfQueue;
-		_mt.unlock();
-	}
+	/**
+	 * \brief erase front element
+	 */
+	void pop();
 
-	//метод обмена очередями
-	void swap(MyQueue& q) noexcept
-	{
-		_mt.lock();
-		q._mt.lock();
-		_q.swap(q._q);
-		q._mt.unlock();
-		_mt.unlock();
-	}
+	/**
+	 * \brief function for swapping queues
+	 * \param q MyQueue for swapping
+	 */
+	void swap(MyQueue& q) noexcept;
 
-	//сетод пытающий удалить первый элемент с последующим его удалением
-	std::pair<bool, T> pull()
-	{
-		std::pair<bool, T> ans;
-		_mt.lock();
-		if (_q.empty())
-			ans.first = false;
-		else 
-		{
-			ans.first = true;
-			ans.second = _q.front().second;
-			_q.pop();
-			--_sizeOfQueue;
-		}
-		_mt.unlock();
-		return ans;
-	}
-	~MyQueue()
-	{
-		while (!_q.empty())
-			_q.pop();
-		if (_withInterface) 
-		{
-			myInterface::MyShower::getInstance().deleteField(_interfaceId1);
-			myInterface::MyShower::getInstance().deleteField(_interfaceId2);
-		}
-	}
+	/**
+	 * \brief function which trys to erase ans return front element
+	 * \return true and object if there was front element or false and defult object if there wasn't front object
+	 */
+	std::pair<bool, T> pull();
+
+	/**
+	 * \brief default destructor
+	 */
+	~MyQueue();
 };
+
+#include "MyQueueDifintion.hpp"
 
 #endif // !MY_QUEUE_DEF
